Move dampers into their own file
